requirements.dev.txt
coverage==7.2.5
flake8==6.0.0
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
packaging==23.1
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.3.1
pytest-cov==4.0.0
pytest-django==4.5.2
